pyproject.toml
[tool.poetry]
name = "ck-price-monitor-bot"
version = "0.1.0"
description = ""
authors = ["Vlad Pastushenko <vladp@ro.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.21.3"
beautifulsoup4 = "^4.10.0"
python-telegram-bot = "^13.10"
lxml = "^4.7.1"
deta = "^1.0.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
isort = "^5.10.1"
black = "^21.12b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
